Move cache evict method into TaskConfig class
